requirements.txt
fastapi==0.88.0
starlette==0.22.0
python-dotenv==0.21.0
requests==2.28.1
uvicorn==0.20.0
Authlib==0.15.5
itsdangerous==2.1.2
httpx==0.23.1
pydantic==1.10.2
SQLAlchemy==1.4.44
